+++ duckscript_sdk/src/sdk/std/collections/mod.rs
@@ -6,6 +6,7 @@ pub(crate) mod array_length;
 pub(crate) mod array_pop;
 mod array_push;
 mod is_array;
+mod map;
 mod range;
 mod read_properties;
 mod write_properties;
@@ -27,6 +28,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(array_length::create(&package))?;
     commands.set(array_pop::create(&package))?;
     commands.set(is_array::create(&package))?;
+    commands.set(map::create(&package))?;
     commands.set(range::create(&package))?;
     commands.set(read_properties::create(&package))?;
     commands.set(write_properties::create(&package))?;